rename ccip to cciptypes
