+++ pkg/types/cciptypes/commitstore.go
@@ -1,4 +1,4 @@
-package ccip
+package cciptypes
 
 import (
 	"context"
