+++ pkg/types/cciptypes/models.go
@@ -1,4 +1,4 @@
-package ccip
+package cciptypes
 
 import (
 	"time"
