+++ pkg/types/cciptypes/priceregistry.go
@@ -1,4 +1,4 @@
-package ccip
+package cciptypes
 
 import (
 	"context"
